issue-51: circumvent deserialization issues
Static fields are initialized at class loading time. When
WicketWebjars*ResourceReference are added on pages or components,
and these pages or components are deserialized (like at container
startup time), WebjarsVersion$Holder is then initialized.

As wicket context is not available at this time, page
deserialization fails.

Holder attributes are moved to methods, so that wicket related
lookups are done on demand, at a time wicket context is correctly
initialized.

This fix does not handle all deserialization issues as it is
needed to ensure that these new methods are called only when
wicket context is available.

I track theses method calls, and it seems to me that they are
always called when wicket context is available.